refactor: Fixing the Vpc provider in Database copy (#709)
